Web_App/requirements.txt
pandas
altair
streamlit
streamlit-authenticator
Image
Pillow
watchdog
pickle
os
sklearn
warnings